Enhance Runar framework with data replication feature and update dependencies
- Introduced a new `data-replication.md` specification outlining requirements for data replication in SQLite services, improving developer experience by automating cross-node synchronization and conflict resolution.
- Updated `Cargo.toml` files to include new dependencies for `runar_services` and `runar_serializer`, ensuring compatibility with the latest features.
- Refactored test utilities to support mobile simulation and database interactions, enhancing the testing framework for end-to-end scenarios.
- Implemented a comprehensive mobile simulator in `runar-test-utils`, allowing for realistic testing of mobile behaviors and interactions with backend services.
- Updated example projects to demonstrate the new data replication capabilities and improved service interactions, ensuring clarity and usability for developers.

+++ rust-examples/micro_services_demo/src/models.rs
@@ -11,10 +11,10 @@ pub struct User {
     #[runar(user, system, search)]
     pub email: String,
 
-    #[runar(user)]
+    #[runar(user, system)]
     pub password_hash: String,
 
-    #[runar(user, system, search)]
+    #[runar(user, system)]
     pub created_at: u64,
 }
 
@@ -28,8 +28,8 @@ pub struct Profile {
     #[runar(user, system, search)]
     pub full_name: String,
 
-    #[runar(user, system, search)]
-    pub bio: String, // Changed from Option<String> to String
+    #[runar(user)]
+    pub bio: String,
 
     #[runar(user)]
     pub private_notes: String,
@@ -59,17 +59,15 @@ pub struct Account {
 pub struct Order {
     pub id: String,
 
-    #[runar(user, system, search)]
     pub user_id: String,
 
-    #[runar(user, system, search)]
     pub product_id: String,
 
-    #[runar(user, system, search)]
+    #[runar(user, system)]
     pub quantity: u32,
 
     #[runar(user)]
-    pub total_price_cents: u64, // Changed from f64 to u64 (price in cents)
+    pub total_price_cents: f64,
 
     #[runar(user, system, search)]
     pub status: String,
